python-worker/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
structlog==24.4.0
openai==1.55.0
redis==5.2.1
httpx==0.27.2
celery==5.4.0
python-dateutil==2.9.0.post0
